# docs/swagger.yaml
info:
  contact: {}
paths:
  /attach/upload:
    post:
      responses: {}
      summary: 发送图片
      tags:
      - 信息模块
  /contact/LoadCommunity:
    post:
      parameters:
      - description: 用户id
        in: query
        name: ownerId
        type: string
      - description: 群名称
        in: query
        name: name
        type: string
      responses:
        "200":
          description: code","msg","data“}
          schema:
            type: string
      summary: 加载群列表
      tags:
      - 群模块
  /contact/createCommunity:
    post:
      parameters:
      - description: 用户id
        in: query
        name: ownerId
        type: string
      - description: 群名称
        in: query
        name: name
        type: string
      responses:
        "200":
          description: code","msg","data“}
          schema:
            type: string
      summary: 新建群
      tags:
      - 群模块
  /deleteRedisMsg:
    post:
      parameters:
      - description: 用户A_id
        in: query
        name: userIdA
        type: string
      - description: 用户B_id
        in: query
        name: userIdB
        type: string
      - description: 删除记录数目
        in: query
        name: nums
        type: string
      responses:
        "200":
          description: code","msg","data“}
          schema:
            type: string
      summary: 读取Redis信息
      tags:
      - 信息模块
  /getRedisMsg:
    post:
      parameters:
      - description: 用户A_id
        in: query
        name: userIdA
        type: string
      - description: 用户B_id
        in: query
        name: userIdB
        type: string
      responses:
        "200":
          description: code","msg","data“}
          schema:
            type: string
      summary: 读取Redis信息
      tags:
      - 信息模块
  /index:
    get:
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - 首页
  /msg/sendMsg:
    get:
      responses: {}
      summary: 发送信息
      tags:
      - 信息模块
  /msg/sendUserMsg:
    get:
      responses: {}
      summary: 发送用户信息
      tags:
      - 信息模块
  /toChat:
    get:
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - 首页
  /toRegister:
    get:
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - 首页
  /user/AddFriends:
    post:
      parameters:
      - description: 用户id
        in: query
        name: userId
        type: string
      - description: 添加用户id
        in: query
        name: targetId
        type: string
      responses:
        "200":
          description: code","msg","data“}
          schema:
            type: string
      summary: 添加好友
      tags:
      - 用户模块
  /user/createUser:
    post:
      parameters:
      - description: 用户名
        in: query
        name: name
        type: string
      - description: 密码
        in: query
        name: password
        type: string
      - description: 确认密码
        in: query
        name: Identity
        type: string
      responses:
        "200":
          description: code","msg","data“}
          schema:
            type: string
      summary: 新增用户
      tags:
      - 用户模块
  /user/deleteUser:
    post:
      parameters:
      - description: id
        in: query
        name: id
        type: string
      responses:
        "200":
          description: code","msg","data“}
          schema:
            type: string
      summary: 删除用户
      tags:
      - 用户模块
  /user/getUserList:
    get:
      responses:
        "200":
          description: code","msg","data“}
          schema:
            type: string
      summary: 获取所有用户信息
      tags:
      - 用户模块
  /user/loginByNameAndPwd:
    post:
      parameters:
      - description: 用户名
        in: query
        name: name
        type: string
      - description: 密码
        in: query
        name: password
        type: string
      - description: 确认密码
        in: query
        name: repassword
        type: string
      responses:
        "200":
          description: code","msg","data“}
          schema:
            type: string
      summary: 用户登录
      tags:
      - 用户模块
  /user/searchFriends:
    post:
      parameters:
      - description: 用户id
        in: query
        name: userId
        type: string
      responses:
        "200":
          description: code","msg","data“}
          schema:
            type: string
      summary: 查询好友列表
      tags:
      - 用户模块
  /user/updateUser:
    post:
      parameters:
      - description: id
        in: formData
        name: id
        type: string
      - description: name
        in: formData
        name: name
        type: string
      - description: password
        in: formData
        name: password
        type: string
      - description: phone
        in: formData
        name: phone
        type: string
      - description: email
        in: formData
        name: email
        type: string
      responses:
        "200":
          description: code","msg","data“}
          schema:
            type: string
      summary: 修改用户
      tags:
      - 用户模块
swagger: "2.0"